requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
python-decouple==3.8
Pillow
celery
redis
stripe
paypal-checkout-serversdk
razorpay
python-dotenv
requests
cryptography
django-filter
django-crispy-forms
crispy-bootstrap5
whitenoise
gunicorn

django-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
stripe==7.8.0
paypal-checkout-serversdk==1.0.1
razorpay==1.4.1
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
gunicorn==21.2.0
